requirements.txt
discord.py >=1.3.4
psutil >=5.6.2
typing >=3.7.4
jishaku >=1.16.3.159
lru-dict==1.1.6
discord-flags==2.1.1
asyncpg==0.20.1
pyxivapi==0.2.0
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@6f2b873bf0d28903eb752aa1166b3aac26dc9007
beautifulsoup4==4.9.1